+++ docs/changelog.rst
@@ -28,6 +28,12 @@ v0.4 (not yet released)
 * Function properties: ``chow_parameters``, ``is_canalizing``, ``is_horn``, ``is_krom``
   `#48 <https://github.com/msoeken/kitty/pull/48>`_
 
+* Spectral canonization with resource limit
+  `#49 <https://github.com/msoeken/kitty/pull/49>`_
+
+* Compute different spectra: ``rademacher_walsh_spectrum``, ``autocorrelation_spectrum``, ``spectrum_distribution``
+  `#49 <https://github.com/msoeken/kitty/pull/49>`_
+
 v0.3 (February 25, 2018)
 ------------------------